refactor: move Popup sign transaction component to its own file SignTransactionFormContent
